# apps/server/swagger.yaml
openapi: 3.0.3
info:
  title: trip_vista_api
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:5000
    description: Local server
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register user
      description: Register user
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user3@gmail.com
                name:
                  type: string
                  example: hezron
                password:
                  type: number
                  example: 123456
            examples:
              Register user:
                value:
                  email: user3@gmail.com
                  name: hezron
                  password: 123456
      responses:
        '200':
          description: ''
...